refactor: AppTarget import 코드 및 테스트 수정
- 불 필요한 import 주입 제거
- 객체별 테스트 형태 -> BDD 형태로 변환 시도
   - 테스트 파일 삭제 CasualConversationTests.swift
- BDD 문맥 한글화
