--- CasualConversation/Targets/CasualConversation/Tests/CasualConversationTests.swift
@@ -1,10 +1,0 @@
-import Foundation
-import XCTest
-
-final class CasualConversationTests: XCTestCase {
-	
-    func test_twoPlusTwo_isFour() {
-        XCTAssertEqual(2+2, 4)
-    }
-	
-}
